fixup: move cli-specific webpack hacks to cli webpack config
